+++ src/classes/scripts.ts
@@ -34,12 +34,7 @@ import {
   RedisJobOptions,
 } from '../types';
 import { ErrorCode } from '../enums';
-import {
-  array2obj,
-  finishedErrors,
-  getParentKey,
-  isRedisVersionLowerThan,
-} from '../utils';
+import { array2obj, getParentKey, isRedisVersionLowerThan } from '../utils';
 import { ChainableCommander } from 'ioredis';
 
 export type JobData = [JobJsonRaw | number, string?];
@@ -223,7 +218,7 @@ export class Scripts {
 
     if (parentOpts.waitChildrenKey) {
       result = await this.addParentJob(client, job, encodedOpts, args);
-    } else if (opts.delay) {
+    } else if (typeof opts.delay == 'number') {
       result = await this.addDelayedJob(client, job, encodedOpts, args);
     } else if (opts.priority) {
       result = await this.addPrioritizedJob(client, job, encodedOpts, args);
@@ -232,7 +227,7 @@ export class Scripts {
     }
 
     if (<number>result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: <number>result,
         parentKey: parentOpts.parentKey,
         command: 'addJob',
@@ -410,21 +405,32 @@ export class Scripts {
     return (<any>client).removeJobScheduler(keys.concat(args));
   }
 
-  async remove(jobId: string, removeChildren: boolean): Promise<number> {
-    const client = await this.queue.client;
-
+  protected removeArgs(jobId: string, removeChildren: boolean): (string | number)[] {
     const keys: (string | number)[] = ['', 'meta'].map(name =>
       this.queue.toKey(name),
     );
-    const result = await (<any>client).removeJob(
-      keys.concat([jobId, removeChildren ? 1 : 0]),
+    
+    const args = [jobId, removeChildren ? 1 : 0];
+
+    return keys.concat(args);
+  }
+
+  async remove(jobId: string, removeChildren: boolean): Promise<number> {
+    const client = await this.queue.client;
+
+    const args = this.removeArgs(
+      jobId, removeChildren
     );
 
-    if (result == ErrorCode.JobBelongsToJobScheduler) {
-      throw finishedErrors({
-        code: ErrorCode.JobBelongsToJobScheduler,
+    const result = await (<any>client).removeJob(
+      args,
+    );
+
+    if (result < 0) {
+      throw this.finishedErrors({
+        code: result,
         jobId,
-        command: 'remove',
+        command: 'removeJob',
       });
     }
 
@@ -460,7 +466,7 @@ export class Scripts {
     const result = await (<any>client).updateData(keys.concat([dataJson]));
 
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId: job.id,
         command: 'updateData',
@@ -486,7 +492,7 @@ export class Scripts {
     );
 
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId,
         command: 'updateProgress',
@@ -511,7 +517,7 @@ export class Scripts {
     );
 
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId,
         command: 'addLog',
@@ -592,7 +598,7 @@ export class Scripts {
 
     const result = await (<any>client).moveToFinished(args);
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId,
         command: 'moveToFinished',
@@ -602,6 +608,49 @@ export class Scripts {
       if (typeof result !== 'undefined') {
         return raw2NextJobData(result);
       }
+    }
+  }
+
+  finishedErrors({
+    code,
+    jobId,
+    parentKey,
+    command,
+    state,
+  }: {
+    code: number;
+    jobId?: string;
+    parentKey?: string;
+    command: string;
+    state?: string;
+  }): Error {
+    switch (code) {
+      case ErrorCode.JobNotExist:
+        return new Error(`Missing key for job ${jobId}. ${command}`);
+      case ErrorCode.JobLockNotExist:
+        return new Error(`Missing lock for job ${jobId}. ${command}`);
+      case ErrorCode.JobNotInState:
+        return new Error(
+          `Job ${jobId} is not in the ${state} state. ${command}`,
+        );
+      case ErrorCode.JobPendingDependencies:
+        return new Error(`Job ${jobId} has pending dependencies. ${command}`);
+      case ErrorCode.ParentJobNotExist:
+        return new Error(`Missing key for parent job ${parentKey}. ${command}`);
+      case ErrorCode.JobLockMismatch:
+        return new Error(
+          `Lock mismatch for job ${jobId}. Cmd ${command} from ${state}`,
+        );
+      case ErrorCode.ParentJobCannotBeReplaced:
+        return new Error(
+          `The parent job ${parentKey} cannot be replaced. ${command}`,
+        );
+      case ErrorCode.JobBelongsToJobScheduler:
+        return new Error(
+          `Job ${jobId} belongs to a job scheduler and cannot be removed directly. ${command}`,
+        );
+      default:
+        return new Error(`Unknown code ${code} error for ${jobId}. ${command}`);
     }
   }
 
@@ -679,7 +728,7 @@ export class Scripts {
       case 1:
         return false;
       default:
-        throw finishedErrors({
+        throw this.finishedErrors({
           code: result,
           jobId,
           parentKey,
@@ -843,7 +892,7 @@ export class Scripts {
     const args = this.changeDelayArgs(jobId, delay);
     const result = await (<any>client).changeDelay(args);
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId,
         command: 'changeDelay',
@@ -880,7 +929,7 @@ export class Scripts {
     const args = this.changePriorityArgs(jobId, priority, lifo);
     const result = await (<any>client).changePriority(args);
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId,
         command: 'changePriority',
@@ -998,7 +1047,7 @@ export class Scripts {
     const args = this.moveToDelayedArgs(jobId, timestamp, token, delay, opts);
     const result = await (<any>client).moveToDelayed(args);
     if (result < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code: result,
         jobId,
         command: 'moveToDelayed',
@@ -1034,7 +1083,7 @@ export class Scripts {
       case 1:
         return false;
       default:
-        throw finishedErrors({
+        throw this.finishedErrors({
           code: result,
           jobId,
           command: 'moveToWaitingChildren',
@@ -1193,7 +1242,7 @@ export class Scripts {
       case 1:
         return;
       default:
-        throw finishedErrors({
+        throw this.finishedErrors({
           code: result,
           jobId: job.id,
           command: 'reprocessJob',
@@ -1257,7 +1306,7 @@ export class Scripts {
 
     const code = await (<any>client).promote(keys.concat(args));
     if (code < 0) {
-      throw finishedErrors({
+      throw this.finishedErrors({
         code,
         jobId,
         command: 'promote',